+++ src/aui-form-validator/js/aui-form-validator.js
@@ -531,6 +531,8 @@ var FormValidator = A.Component.create({
 
             instance.errors = {};
             instance._blurHandlers = null;
+            instance._fileBlurHandlers = null;
+            instance._fileInputHandlers = null;
             instance._inputHandlers = null;
             instance._rulesAlreadyExtracted = false;
             instance._stackErrorContainers = {};
@@ -995,19 +997,6 @@ var FormValidator = A.Component.create({
         },
 
         /**
-         * Fires after `validateOnInput` attribute change.
-         *
-         * @method _afterValidateOnInputChange
-         * @param event
-         * @protected
-         */
-        _afterValidateOnInputChange: function(event) {
-            var instance = this;
-
-            instance._uiSetValidateOnInput(event.newVal);
-        },
-
-        /**
          * Fires after `validateOnBlur` attribute change.
          *
          * @method _afterValidateOnBlurChange
@@ -1018,6 +1007,19 @@ var FormValidator = A.Component.create({
             var instance = this;
 
             instance._uiSetValidateOnBlur(event.newVal);
+        },
+
+        /**
+         * Fires after `validateOnInput` attribute change.
+         *
+         * @method _afterValidateOnInputChange
+         * @param event
+         * @protected
+         */
+        _afterValidateOnInputChange: function(event) {
+            var instance = this;
+
+            instance._uiSetValidateOnInput(event.newVal);
         },
 
         /**
@@ -1339,12 +1341,21 @@ var FormValidator = A.Component.create({
             if (val) {
                 if (!instance._inputHandlers) {
                     instance._inputHandlers = boundingBox.delegate('input', instance._onFieldInput,
-                        'input,select,textarea,button', instance);
+                        'input:not([type="file"]),select,textarea,button', instance);
+                }
+
+                if (!instance._fileInputHandlers) {
+                    instance._fileInputHandlers = boundingBox.delegate('change', instance._onFieldInput,
+                        'input[type="file"]', instance);
                 }
             }
             else {
                 if (instance._inputHandlers) {
                     instance._inputHandlers.detach();
+                }
+
+                if (instance._fileInputHandlers) {
+                    instance._fileInputHandlers.detach();
                 }
             }
         },
@@ -1363,12 +1374,21 @@ var FormValidator = A.Component.create({
             if (val) {
                 if (!instance._blurHandlers) {
                     instance._blurHandlers = boundingBox.delegate('blur', instance._onFieldInput,
-                        'input,select,textarea,button', instance);
+                        'input:not([type="file"]),select,textarea,button', instance);
+                }
+
+                if (!instance._fileBlurHandlers) {
+                    instance._fileBlurHandlers = boundingBox.delegate('change', instance._onFieldInput,
+                        'input[type="file"]', instance);
                 }
             }
             else {
                 if (instance._blurHandlers) {
                     instance._blurHandlers.detach();
+                }
+
+                if (instance._fileBlurHandlers) {
+                    instance._fileBlurHandlers.detach();
                 }
             }
         }