+++ src/aui-layout/HISTORY.md
@@ -5,6 +5,7 @@
 ## @VERSION@
 
 * [AUI-2030](https://issues.liferay.com/browse/AUI-2030) A row with one col should be automatically added on last position when the last row have at least one field
+* [AUI-2036](https://issues.liferay.com/browse/AUI-2036) AUI Form Builder Field Settings Modal should not overwrite Bootstrap 3 Modal css rules so that the layout keeps as similar as possible
 * [AUI-2031](https://issues.liferay.com/browse/AUI-2031) JavaScript error thrown when trying to split a row into two columns
 * [AUI-2016](https://issues.liferay.com/browse/AUI-2016) Remove column delete button from From Builder
 * [AUI-1930](https://issues.liferay.com/browse/AUI-1930) Change functionality of add cols